+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -5,10 +5,11 @@ CHANGELOG
 -----
 
  * Started using ICU parent locales as fallback locales.
+ * allow using the ICU message format using domains with the "+intl-icu" suffix
  * deprecated `Translator::transChoice()` in favor of using `Translator::trans()` with a `%count%` parameter
  * deprecated `TranslatorInterface` in favor of `Symfony\Contracts\Translation\TranslatorInterface`
  * deprecated `MessageSelector`, `Interval` and `PluralizationRules`; use `IdentityTranslator` instead
- * Added `IntlMessageFormatter` and `FallbackMessageFormatter`
+ * Added `IntlFormatter` and `IntlFormatterInterface`
  * added support for multiple files and directories in `XliffLintCommand`
  * Marked `Translator::getFallbackLocales()` and `TranslationDataCollector::getFallbackLocales()` as internal